requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes
-e .

alembic==1.17.1
    # via pdf2audiobook
amqp==5.3.1
    # via kombu
annotated-doc==0.0.3
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
azure-cognitiveservices-speech==1.46.0
    # via pdf2audiobook
azure-core==1.36.0
    # via azure-cognitiveservices-speech
bcrypt==5.0.0
    # via passlib
billiard==4.2.2
    # via celery
boto3==1.40.64
    # via pdf2audiobook
botocore==1.40.64
    # via
    #   boto3
    #   s3transfer
cachetools==6.2.1
    # via google-auth
celery==5.5.3
    # via pdf2audiobook
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0 ; platform_python_implementation != 'PyPy'
    # via cryptography
charset-normalizer==3.4.4
    # via
    #   reportlab
    #   requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   uvicorn
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6 ; sys_platform == 'win32'
    # via
    #   click
    #   loguru
    #   pytest
    #   tqdm
    #   uvicorn
cryptography==46.0.3
    # via python-jose
deprecated==1.3.1
    # via limits
distro==1.9.0
    # via openai
dnspython==2.8.0
    # via email-validator
ecdsa==0.19.1
    # via python-jose
elevenlabs==2.21.0
    # via pdf2audiobook
email-validator==2.3.0
    # via pydantic
fastapi==0.121.0
    # via pdf2audiobook
google-api-core==2.28.1
    # via google-cloud-texttospeech
google-auth==2.42.1
    # via
    #   google-api-core
    #   google-cloud-texttospeech
google-cloud-texttospeech==2.33.0
    # via pdf2audiobook
googleapis-common-protos==1.71.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.2.4 ; platform_machine == 'AMD64' or platform_machine == 'WIN32' or platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'ppc64le' or platform_machine == 'win32' or platform_machine == 'x86_64'
    # via sqlalchemy
grpcio==1.76.0
    # via
    #   google-api-core
    #   google-cloud-texttospeech
    #   grpcio-status
grpcio-status==1.76.0
    # via google-api-core
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   elevenlabs
    #   openai
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
jiter==0.11.1
    # via openai
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kombu==5.5.4
    # via celery
limits==5.6.0
    # via slowapi
loguru==0.7.3
    # via pdf2audiobook
mako==1.3.10
    # via alembic
markupsafe==3.0.3
    # via mako
openai==2.6.1
    # via pdf2audiobook
packaging==25.0
    # via
    #   kombu
    #   limits
    #   pytesseract
    #   pytest
passlib==1.7.4
    # via pdf2audiobook
pdf2image==1.17.0
    # via pdf2audiobook
pillow==12.0.0
    # via
    #   pdf2image
    #   pytesseract
    #   reportlab
pluggy==1.6.0
    # via pytest
prompt-toolkit==3.0.52
    # via click-repl
proto-plus==1.26.1
    # via
    #   google-api-core
    #   google-cloud-texttospeech
protobuf==6.33.0
    # via
    #   google-api-core
    #   google-cloud-texttospeech
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
psycopg2-binary==2.9.11
    # via pdf2audiobook
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   python-jose
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pycparser==2.23 ; implementation_name != 'PyPy' and platform_python_implementation != 'PyPy'
    # via cffi
pydantic==2.12.3
    # via
    #   elevenlabs
    #   fastapi
    #   openai
    #   pdf2audiobook
    #   pydantic-settings
pydantic-core==2.41.4
    # via
    #   elevenlabs
    #   pydantic
pydantic-settings==2.11.0
    # via pdf2audiobook
pydub==0.25.1
    # via pdf2audiobook
pygments==2.19.2
    # via pytest
pymupdf==1.26.5
    # via pdf2audiobook
pytesseract==0.3.13
    # via pdf2audiobook
pytest==8.4.2
    # via pytest-asyncio
pytest-asyncio==1.2.0
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   celery
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose==3.5.0
    # via pdf2audiobook
python-multipart==0.0.20
    # via pdf2audiobook
pyyaml==6.0.3
    # via uvicorn
redis==7.0.1
    # via pdf2audiobook
reportlab==4.4.4
    # via pdf2audiobook
requests==2.32.5
    # via
    #   azure-core
    #   elevenlabs
    #   google-api-core
rsa==4.9.1
    # via
    #   google-auth
    #   python-jose
s3transfer==0.14.0
    # via boto3
six==1.17.0
    # via
    #   ecdsa
    #   python-dateutil
slowapi==0.1.9
    # via pdf2audiobook
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.44
    # via
    #   alembic
    #   pdf2audiobook
starlette==0.49.3
    # via fastapi
tqdm==4.67.1
    # via openai
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   azure-core
    #   elevenlabs
    #   fastapi
    #   grpcio
    #   limits
    #   openai
    #   pydantic
    #   pydantic-core
    #   pytest-asyncio
    #   sqlalchemy
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via
    #   botocore
    #   requests
uvicorn==0.38.0
    # via pdf2audiobook
uvloop==0.22.1 ; platform_python_implementation != 'PyPy' and sys_platform != 'cygwin' and sys_platform != 'win32'
    # via uvicorn
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
watchfiles==1.1.1
    # via uvicorn
wcwidth==0.2.14
    # via prompt-toolkit
websockets==15.0.1
    # via
    #   elevenlabs
    #   uvicorn
win32-setctime==1.2.0 ; sys_platform == 'win32'
    # via loguru
wrapt==2.0.0
    # via deprecated
